+++ src/BIT.Data.Sync/EventArgs/NodeSavedDeltaEventArgs.cs
@@ -1,0 +1,17 @@
+﻿using BIT.Data.Sync.Server;
+using System;
+
+namespace BIT.Data.Sync.EventArgs
+{
+    public class NodeSavedDeltaEventArgs : SaveDeltaBaseEventArgs
+    {
+
+        public NodeSavedDeltaEventArgs(IDelta delta, ISyncServerNode node, SaveDeltaBaseEventArgs saveDeltaBaseEventArgs) : base(delta)
+        {
+            Node = node;
+            this.saveDeltaBaseEventArgs = saveDeltaBaseEventArgs;
+        }
+        SaveDeltaBaseEventArgs saveDeltaBaseEventArgs { get; set; }
+        public ISyncServerNode Node { get; set; }
+    }
+}