+++ src/BIT.Data.Sync/EventArgs/ProcessDeltaBaseEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace BIT.Data.Sync.EventArgs
+{
+    public class ProcessDeltaBaseEventArgs : SyncEventBase
+    {
+        public IDelta Delta { get; set; }
+
+        public ProcessDeltaBaseEventArgs(IDelta delta)
+        {
+            Delta = delta;
+        }
+    }
+}