+++ src/BIT.Data.Sync/EventArgs/SavedDeltaEventArgs.cs
@@ -1,0 +1,14 @@
+﻿using System;
+
+namespace BIT.Data.Sync.EventArgs
+{
+    public class SavedDeltaEventArgs : SaveDeltaBaseEventArgs
+    {
+        public SavedDeltaEventArgs(IDelta delta) : base(delta)
+        {
+
+        }
+
+
+    }
+}