+++ src/BIT.Data.Sync/EventArgs/SyncEventBase.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+
+namespace BIT.Data.Sync.EventArgs
+{
+    public abstract class SyncEventBase : EventArgs
+    {
+
+        protected SyncEventBase()
+        {
+
+            Data = new Dictionary<string, object>();
+        }
+        public Dictionary<string, object> Data { get; set; }
+    }
+}