+++ src/BIT.Data.Sync/IDeltaProcessorWithEvents.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using BIT.Data.Sync.EventArgs;
 
 namespace BIT.Data.Sync
 {