+++ src/BIT.Data.Sync/Server/SyncServer.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using BIT.Data.Sync.EventArgs;
 
 namespace BIT.Data.Sync.Server
 {